+++ src/tests/markdownit/mentions.spec.js
@@ -1,0 +1,13 @@
+import markdownit from '../../markdownit'
+import stripIndent from './stripIndent.js'
+
+describe('markdownit', () => {
+
+	it('renders mentions of users with escaped whitespace', () => {
+		const rendered = markdownit.render('@[whitespace user](mention://user/whitespace%20user)')
+		expect(stripIndent(rendered)).toBe(stripIndent(`
+			<p><span class="mention" data-type="user" data-id="whitespace%20user">whitespace user</span></p>`
+		))
+	})
+
+})
