+++ src/tests/markdownit/stripIndent.js
@@ -1,0 +1,7 @@
+export default function stripIndent(content) {
+	return content
+		.replace(/\n/g, "")
+		.replace(/[\t ]+\</g, "<")
+		.replace(/\>[\t ]+\</g, "><")
+		.replace(/\>[\t ]+$/g, ">")
+}
